+++ langserve/playground/src/utils/str.ts
@@ -1,0 +1,5 @@
+export function str(o: unknown): React.ReactNode {
+  return typeof o === "object"
+    ? JSON.stringify(o, null, 2)
+    : (o as React.ReactNode);
+}
